Adjust file structure
